Delete cinema

+++ src/main/java/dat3/voximovies/service/CinemaService.java
@@ -3,9 +3,15 @@ package dat3.voximovies.service;
 import dat3.voximovies.dto.CinemaRequest;
 import dat3.voximovies.dto.CinemaResponse;
 import dat3.voximovies.entity.Cinema;
+
+import dat3.voximovies.entity.Showing;
+import dat3.voximovies.entity.User;
+import dat3.voximovies.repository.*;
+
 import dat3.voximovies.entity.User;
 import dat3.voximovies.repository.CinemaRepository;
 import dat3.voximovies.repository.UserRepository;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -20,12 +26,22 @@ import java.util.Optional;
 public class CinemaService {
 
     CinemaRepository cinemaRepository;
+
+    ReservationRepository reservationRepository;
+    ReviewRepository reviewRepository;
+    ShowingRepository showingRepository;
     UserRepository userRepository;
 
-    public CinemaService(CinemaRepository cinemaRepository, UserRepository userRepository){
-        this.cinemaRepository=cinemaRepository;
-        this.userRepository=userRepository;
+    public CinemaService(CinemaRepository cinemaRepository, ReservationRepository reservationRepository, ReviewRepository reviewRepository, ShowingRepository showingRepository, UserRepository userRepository) {
+        this.cinemaRepository = cinemaRepository;
+        this.reservationRepository = reservationRepository;
+        this.reviewRepository = reviewRepository;
+        this.showingRepository = showingRepository;
+        this.userRepository = userRepository;
+
+ 
     }
+
     public Cinema findCinemaByID(Long id) {
         Cinema cinema = cinemaRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Cinema with this ID doesnt exist"));
         return cinema;
@@ -48,9 +64,17 @@ public class CinemaService {
         return new CinemaResponse(cinema);
     }
 
-    public void deleteCinema(Long id) {
-        Cinema cinema = findCinemaByID(id);
-        cinemaRepository.delete(cinema);
+    public void deleteCinema(Long id, String owner) {
+        User user = userRepository.findByUsername(owner);
+        if(cinemaRepository.existsCinemaByOwner(user)){
+            Cinema cinema = findCinemaByID(id);
+            ArrayList<Showing> showings = showingRepository.findAllByCinemaId(cinema.getId());
+            long showingID = showings.get(0).getId();
+            reservationRepository.deleteAllByShowingId(showingID);
+            showingRepository.deleteAllByCinema(cinema);
+            cinemaRepository.delete(cinema);
+        }
+
     }
 
     public CinemaResponse editCinema(Long id, CinemaRequest request) {
